简述：调试线下整合接口
背景信息：

修改内容：

+++ app/src/main/kotlin/com/liking/treadmill/socket/data/result/UserInfoResultData.kt
@@ -41,6 +41,6 @@ class UserInfoResultData : ResultData() {
         @SerializedName("Gender")
         var gender: Int = 0
         @SerializedName("BraceletId")
-        var braceletId: Int = 0
+        var braceletId: Long = 0
     }
 }
